update extraction script to work with the Phillips scans
This was my first attempt, but actually it's not quite right.
Because I am extracting each archive file into its own directory
(as opposed to extracting them all to the same target directory)
the phillips scans won't end up as two images in the same directory
as the current change assumes. This will be fixed in a future commit.